algorithm/src/test/java/org/example/inflearn/Q0810Test.java
package org.example.inflearn;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;

class Q0810Test {

    @Test
    void solution() {
        Q0810 sut = new Q0810();
        int expected;
        int actual;

        expected = 8;
        actual = sut.solution(new int[][]{
            {0, 0, 0, 0, 0, 0, 0},
            {0, 1, 1, 1, 1, 1, 0},
            {0, 0, 0, 1, 0, 0, 0},
            {1, 1, 0, 1, 0, 1, 1},
            {1, 1, 0, 0, 0, 0, 1},
            {1, 1, 0, 1, 1, 0, 0},
            {1, 0, 0, 0, 0, 0, 0},
        });

        assertThat(actual).isEqualTo(expected);

    }
}